+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 7.56
 -----
+- Allows navigation from the Player to the Podcast screen by tapping on Podcast name [#1322]
 
 7.55
 -----